# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: HungerHunt Warehouse–Accounts API
  version: 1.0.0
  description: Versioned contract for procurement review and deterministic inventory analytics.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /purchase-orders:
    post:
      summary: Submit an inventory purchase order for Accounts review
      description: Allowed roles are admin and warehouse.
      operationId: createPurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePurchaseOrderRequest' }
      responses:
        '201':
          description: Created in PENDING_REVIEW
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseOrderEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      summary: List versioned purchase orders
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/PurchaseOrderStatus' }
      responses:
        '200':
          description: Purchase orders, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PurchaseOrder' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /analytics/inventory:
    get:
      summary: Get dashboard- and LLM-ready deterministic analytics
      description: Read-only. It never mutates reorder settings or submits an order.
      operationId: getInventoryAnalytics
      parameters:
        - name: asOf
          in: query
          description: Optional ISO-8601 effective time, useful for reproducible reports.
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Versioned analytics report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalyticsEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /fulfillment-orders:
    get:
      summary: List dorm packages awaiting fulfilment
      description: Admin and warehouse roles. Defaults to active package states, ordered by delivery deadline.
      operationId: listFulfillmentOrders
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/FulfillmentOrderStatus' }
      responses:
        '200':
          description: At most 500 packages
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/FulfillmentOrder' } }
                  meta:
                    allOf:
                      - { $ref: '#/components/schemas/ListMeta' }
                      - { $ref: '#/components/schemas/RoomUnitsMeta' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /replenishment-drafts:
    post:
      summary: Generate or return today's active analytics-assisted replenishment draft
      description: Warehouse/admin only. Open purchase-order quantities are subtracted and the draft expires after seven days.
      operationId: generateReplenishmentDraft
      responses:
        '200': { description: Existing active draft returned }
        '201': { description: Draft created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /caretaker/reports:
    post:
      summary: Raise a package issue or a professional complaint
      description: >-
        A caretaker's channel to the school office. Two kinds share one record: ORDER_ISSUE is
        about one package and carries orderId, scoped to the caretaker's own rooms exactly as
        the collection route is (a package in a room they do not hold answers 404, never a
        refusal, so this cannot be used to learn one exists); COMPLAINT is anything else,
        including about the warehouse or another member of staff. Nothing about who is
        reporting is read from the body — the name, role and rooms are copied onto the report
        from the roster and the
        session. Raising an issue never changes the package: a reported package stays
        collectable, because a student owed food should not lose it while an office reads a
        message.
      operationId: createStaffReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [kind, category, note]
              properties:
                kind: { $ref: '#/components/schemas/StaffReportKind' }
                category:
                  type: string
                  description: >-
                    One of the fixed categories for this kind. ORDER_ISSUE: MISSING_ITEM,
                    WRONG_ITEM, DAMAGED, NOT_THEIR_ORDER, CODE_NOT_WORKING, OTHER. COMPLAINT:
                    WORKING_CONDITIONS, STAFF_CONDUCT, STUDENT_WELFARE, DELIVERY_SERVICE,
                    APP_PROBLEM, OTHER.
                note: { type: string, minLength: 10, maxLength: 1000 }
                orderId:
                  type: string
                  pattern: '^[a-fA-F0-9]{24}$'
                  description: Required for ORDER_ISSUE, ignored for COMPLAINT.
      responses:
        '201':
          description: Report raised
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/StaffReport' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429':
          description: This caretaker already has 25 reports the office has not answered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    get:
      summary: The caretaker's own reports and what came back
      description: >-
        Filtered to the account in the session; there is no query that widens it. The reply the
        office wrote is included, and the name of the admin who wrote it — every admin sees the
        office queue and any may answer any report, so the answer carries its author. The
        internal handling trail stays with the office.
      operationId: listOwnStaffReports
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: The caretaker's reports, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/StaffReport' } }
                  meta: { $ref: '#/components/schemas/Meta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports:
    get:
      summary: The office queue of caretaker reports
      description: >-
        Every admin account reaches the same queue; there is no owner and no assignment.
        Deliberately not open to warehouse or caretaker accounts: a complaint may be about
        either, and a channel its subject can read is not one anybody uses twice. Ordered
        oldest-first while showing what is still owed, newest-first when read as a log — not by
        status, which sorts alphabetically and would rank ACKNOWLEDGED above unread OPEN.
      operationId: listStaffReports
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OUTSTANDING, OPEN, ACKNOWLEDGED, RESOLVED] }
          description: OUTSTANDING means OPEN or ACKNOWLEDGED — everything the office still owes.
        - name: kind
          in: query
          schema: { $ref: '#/components/schemas/StaffReportKind' }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: Reports, with the outstanding count in meta
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/StaffReport' } }
                  meta: { $ref: '#/components/schemas/Meta' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PurchaseOrderStatus:
      type: string
      enum: [PENDING_REVIEW, APPROVED, REJECTED, PARTIALLY_RECEIVED, RECEIVED, CANCELLED]
    StaffReportKind:
      type: string
      enum: [ORDER_ISSUE, COMPLAINT]
    StaffReport:
      type: object
      description: >-
        A report raised by a caretaker. raisedBy and handling are present only in the admin
        console's view; the caretaker reading their own report sees the reply but not who wrote
        it.
      required: [id, kind, category, note, status, raisedAt]
      properties:
        id: { type: string }
        kind: { $ref: '#/components/schemas/StaffReportKind' }
        category: { type: string }
        categoryLabel: { type: string, description: The category in the words the apps show. }
        note: { type: string }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        raisedAt: { type: string, format: date-time }
        acknowledgedAt: { type: [string, 'null'], format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        resolutionNote: { type: string, description: What the office said back. Read by the caretaker. }
        answeredBy:
          type: string
          description: >-
            Name of the admin who resolved it, snapshotted at the time. Shown to the caretaker
            with the reply, because every admin can answer any report and the answer carries its
            author. Empty until the report is resolved.
        order:
          type: [object, 'null']
          description: Present on ORDER_ISSUE. Carries no price — a caretaker never sees what a package cost.
          required: [id, studentName, statusAtReport]
          properties:
            id: { type: string }
            studentName: { type: string }
            roomNumber: { type: string }
            statusAtReport: { $ref: '#/components/schemas/FulfillmentOrderStatus' }
        raisedBy:
          type: object
          description: Admin view only. Copied from the roster when the report was raised.
          properties:
            id: { type: string }
            name: { type: string }
            role: { type: string }
            roomNumbers:
              type: string
              description: "Every room the raiser held when they wrote it, joined with ' · '."
        handling:
          type: array
          description: Admin view only. Append-only trail of who read and answered it.
          items: { type: object }
    FulfillmentOrderStatus:
      type: string
      enum: [PENDING, PACKED, OUT_FOR_DELIVERY, DELIVERED, COLLECTED, CANCELLED]
    FulfillmentOrder:
      type: object
      required: [id, transactionId, studentId, student, items, totalAmount, status, businessWeekStart, orderedAt, deliverBy]
      properties:
        id: { type: string }
        transactionId: { type: string }
        studentId: { type: string }
        student:
          type: object
          required: [name, roomNumber]
          properties:
            name: { type: string }
            admissionNumber: { type: string }
            roomNumber: { type: string }
            roomId:
              type: string
              description: >-
                The room this package belongs to, snapshotted at order time. Load-bearing, not
                decorative: the warehouse board maps every order onto its delivery unit by this
                id alone (never by roomNumber), so dropping it from the snapshot silently
                ungroups the whole board.
        items: { type: array, items: { type: object } }
        totalAmount: { type: number }
        status: { $ref: '#/components/schemas/FulfillmentOrderStatus' }
        businessWeekStart: { type: string, format: date-time }
        orderedAt: { type: string, format: date-time }
        deliverBy: { type: string, format: date-time }
        packedAt: { type: [string, 'null'], format: date-time }
        dispatchedAt: { type: [string, 'null'], format: date-time }
        deliveredAt:
          type: [string, 'null']
          format: date-time
          description: When the warehouse handed the package to the room's caretaker.
        collectedAt:
          type: [string, 'null']
          format: date-time
          description: >-
            When the student took the package from the caretaker, recorded only by that
            student entering their own purchase code. Null while the package is still
            with the caretaker.
        deliveryNote: { type: string }
        proofOfDelivery:
          type: [object, 'null']
          required: [receivedBy, receiverPhone, recordedBy, recordedAt]
          properties:
            receivedBy: { type: string, maxLength: 60 }
            receiverPhone: { type: string, pattern: '^\d{10}$' }
            recordedBy: { type: string, description: The staff account that recorded the handover. }
            recordedAt: { type: string, format: date-time }
...
    CreatePurchaseOrderRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        supplierId: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
        reason: { type: string, maxLength: 500 }
        items:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [productId, quantity]
            properties:
              productId: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
              quantity: { type: integer, minimum: 1 }
              estimatedUnitCost: { type: number, minimum: 0 }
...
    PurchaseOrder:
      type: object
      required: [id, status, items, submittedAt]
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/PurchaseOrderStatus' }
        supplierId: { type: [string, 'null'] }
        reason: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string }
              quantity: { type: integer }
              receivedQuantity: { type: number }
              estimatedUnitCost: { type: number }
        requestedBy: { type: [string, 'null'] }
        reviewedBy: { type: [string, 'null'] }
        reviewReason: { type: string }
        submittedAt: { type: string, format: date-time }
        reviewedAt: { type: [string, 'null'], format: date-time }
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string }
    ListMeta:
      allOf:
        - { $ref: '#/components/schemas/Meta' }
        - type: object
          required: [count]
          properties: { count: { type: integer, minimum: 0 } }
    RoomUnit:
      type: object
      description: >-
        One delivery unit: the set of rooms that travel together as a single trolley, handover
        and report. Rooms covered by an identical set of active caretakers form one unit; a room
        nobody covers stands alone. Derived on read, never stored, and ordered naturally by the
        first room code so the board does not reshuffle between refreshes.
      required: [rooms]
      properties:
        rooms:
          type: array
          minItems: 1
          items:
            type: object
            required: [id, code]
            properties:
              id:
                type: string
                description: Matches FulfillmentOrder.student.roomId — this is how orders join to a unit.
              code: { type: string }
    RoomUnitsMeta:
      type: object
      description: >-
        Staff-only grouping of the board. Absent from caretaker responses, which carry
        awaitingCollection instead — a caretaker already knows their own rooms and everybody
        else's grouping is not theirs to see.
      properties:
        roomUnits:
          type: array
          items: { $ref: '#/components/schemas/RoomUnit' }
    PurchaseOrderEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/PurchaseOrder' }
        meta: { $ref: '#/components/schemas/Meta' }
    AnalyticsEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [schemaVersion, generatedAt, methodology, summary, procurement, items]
          properties:
            schemaVersion: { type: string, const: '1.0' }
            generatedAt: { type: string, format: date-time }
            methodology: { type: object }
            summary: { type: object }
            procurement: { type: object }
            items: { type: array, items: { type: object } }
        meta:
          allOf:
            - { $ref: '#/components/schemas/Meta' }
            - type: object
              properties:
                deterministic: { type: boolean, const: true }
                llmReady: { type: boolean, const: true }
...
    ErrorEnvelope:
      type: object
      description: >-
        The structured v1 error contract. Paths backed by a legacy (pre-v1) controller —
        including every /payments/* path in this document — do not actually produce this shape:
        they send only { message }, with no error object and no meta. Where this schema is
        referenced from such a path, read it as documenting message alone; error and meta will
        not be present on the wire.
      required: [message, error, meta]
      properties:
        message: { type: string, description: Legacy-client compatibility alias. }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: object } }
        meta: { $ref: '#/components/schemas/Meta' }
  responses:
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated role lacks permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }